Add files via upload
This is one of my first jobs - a script to bypass blocking while parsing. This script returns the address of the proxy server + one of the user-agents from the list which is also in this repository. 
Since this is one of my first programs, you can track the difference between this script and my other scripts, progress :)